move data generation functions to data.py
