+++ scripts/charge_benchmark.py
@@ -1,13 +1,10 @@
-#generate gravitational field
 import sys
 import numpy as np
-import itertools as it
 from functools import partial
 import argparse
 import time
 import math
 import matplotlib.pyplot as plt
-import imageio
 
 from jax import vmap, jit
 import jax.numpy as jnp
@@ -18,82 +15,7 @@ import optax
 
 import geometricconvolutions.geometric as geom
 import geometricconvolutions.ml as ml
-import geometricconvolutions.utils as utils
-
-def get_initial_charges(num_charges, N, D, rand_key):
-    return N*random.uniform(rand_key, shape=(num_charges, D))
-
-def get_velocity_vector(loc, charge_loc, charge):
-    vec = loc - charge_loc
-    scaling = jnp.linalg.norm(vec) ** 3
-    return (charge / scaling) * vec
-
-def get_velocity_field(N, D, charges):
-    pixel_idxs = jnp.array(list(it.product(range(N), repeat=D)), dtype=int)
-    velocity_field = jnp.zeros((N,)*D + (D,))
-
-    vmap_get_vv = vmap(get_velocity_vector, in_axes=(0, None, None)) #all locs, one charge
-
-    for charge in charges:
-        velocity_field = velocity_field + vmap_get_vv(pixel_idxs, charge, 1).reshape((N,)*D + (D,))
-
-    return geom.GeometricImage(velocity_field, 0, D, is_torus=False)
-
-def update_charges(charges, delta_t):
-    get_net_velocity = vmap(get_velocity_vector, in_axes=(None, 0, None)) #single loc, all charges
-
-    new_charges = []
-    for i in range(len(charges)):
-        velocities = get_net_velocity(charges[i], jnp.concatenate((charges[:i], charges[i+1:])), 1)
-        assert velocities.shape == (len(charges) - 1, 2)
-        net_velocity = jnp.sum(velocities, axis=0)
-        assert net_velocity.shape == charges[i].shape == (2,)
-        new_charges.append(charges[i] + delta_t * net_velocity)
-    return jnp.stack(new_charges)
-
-def Qtransform(vector_field, s):
-    vector_field_norm = vector_field.norm()
-    return geom.GeometricImage(
-        (4*(jax.nn.sigmoid(vector_field_norm.data / s)-0.5)) / vector_field_norm.data, 
-        0, 
-        vector_field.D,
-        is_torus=vector_field.is_torus,
-    ) * vector_field
-
-def get_data(N, D, num_charges, num_steps, delta_t, s, rand_key, num_images=1, outfile=None, warmup_steps=0):
-    assert (not outfile) or (num_images == 1)
-
-    initial_fields = []
-    final_fields = []
-    for _ in range(num_images):
-        rand_key, subkey = random.split(rand_key)
-        # generate charges, generally in the center so that they don't repel off the grid
-        charges = get_initial_charges(num_charges, N/2, D, subkey) + jnp.array([int(N/4)]*D)
-        for i in range(warmup_steps):
-            charges = update_charges(charges, delta_t)
-
-        initial_fields.append(Qtransform(get_velocity_field(N, D, charges), s))
-        if outfile:
-            utils.plot_image(initial_fields[-1])
-            plt.savefig(f'{outfile}_{0}.png')
-            plt.close()
-
-        for i in range(1,num_steps+1):
-            charges = update_charges(charges, delta_t)
-            if outfile:
-                utils.plot_image(Qtransform(get_velocity_field(N, D, charges), s))
-                plt.savefig(f'{outfile}_{i}.png')
-                plt.close()
-
-        if outfile:
-            with imageio.get_writer(f'{outfile}.gif', mode='I') as writer:
-                for i in range(num_steps+1):
-                    image = imageio.imread(f'{outfile}_{i}.png')
-                    writer.append_data(image)
-
-        final_fields.append(Qtransform(get_velocity_field(N, D, charges), s))
-
-    return initial_fields, final_fields
+from geometricconvolutions.data import get_charge_data as get_data
 
 def net(params, x, D, is_torus, conv_filters, target_img, return_params=False):
     # Note that the target image is only used for its shape
@@ -217,7 +139,6 @@ s = 0.2 # used for transforming the input/output data
 
 num_points = 3
 num_train_images_range = [5,10,20,50,100]
-# num_train_images = 100
 num_test_images = 10
 num_val_images = 10
 
@@ -250,7 +171,7 @@ _, num_params = net(
     one_point_y.data,
     return_params=True,
 )
-print(num_params)
+print('Model Params:', num_params)
 
 _, baseline_num_params = baseline_net(
     huge_params,
@@ -260,7 +181,7 @@ _, baseline_num_params = baseline_net(
     one_point_y.data,
     return_params=True,
 )
-print(baseline_num_params)
+print('Baseline Params:', baseline_num_params)
 models = [
     (
         partial(map_and_loss, D=one_point_x.D, is_torus=one_point_x.is_torus, conv_filters=conv_filters), 
@@ -355,13 +276,11 @@ plt.rcParams['font.serif'] = 'STIXGeneral'
 plt.tight_layout()
 
 plt.plot(num_train_images_range, model_train_loss, label='GI-Net train', color='b', marker='o', linestyle='dashed')
-# plt.plot(num_train_images_range, model_val_loss, label='model val', marker='o')
 plt.plot(num_train_images_range, model_test_loss, label='GI-Net test', color='b', marker='o')
 plt.plot(num_train_images_range, baseline_train_loss, label='Baseline train', color='r', marker='o', linestyle='dashed')
-# plt.plot(num_train_images_range, baseline_val_loss, label='baseline val', marker='o')
 plt.plot(num_train_images_range, baseline_test_loss, label='Baseline test', color='r', marker='o')
 plt.legend()
 plt.title('Charge Field Loss vs. Number of Training Points')
 plt.xlabel('Number of Training Points')
 plt.ylabel('RMSE Loss')
-plt.savefig('../images/charge/charge_loss_chart_1729.png')
+plt.savefig(f'../images/charge/charge_loss_chart_{seed}.png')
