+++ tests/test_misc.py
@@ -83,7 +83,6 @@ class TestMisc:
             assert pair in idxs
 
         idxs = geom.get_contraction_indices(5,1, ((0,1),(2,3)))
-        print(idxs)
         known_list = [
             ((0,1),(2,3)), ((0,1),(2,4)), ((0,2),(1,3)), ((0,2),(1,4)), ((0,2),(3,4)), ((0,4),(2,3)),
         ]
